RFR - DefSigCompleter needs to be merged first
